requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
lxml>=5.0.0
cryptography>=41.0.7
gunicorn>=21.2.0
